renamed CollinsError.js to collins-error.js
