+++ libs/Configuration.js
@@ -1,5 +1,5 @@
 'use strict';
-const CollinsError = require('./CollinsError');
+const CollinsError = require('./collins-error');
 const convict = require('convict');
 const Loader = require('../utils/Loader');
 const Config = require('../configs');
